Move paste functions to their own file
